+++ pkg/transport/process/processors/resource_labeler.go
@@ -11,6 +11,7 @@ import (
 	cdv2 "github.com/gardener/component-spec/bindings-go/apis/v2"
 
 	"github.com/gardener/component-cli/pkg/transport/process"
+	"github.com/gardener/component-cli/pkg/transport/process/utils"
 )
 
 type resourceLabeler struct {
@@ -26,7 +27,7 @@ func NewResourceLabeler(labels ...cdv2.Label) process.ResourceStreamProcessor {
 }
 
 func (p *resourceLabeler) Process(ctx context.Context, r io.Reader, w io.Writer) error {
-	cd, res, resBlobReader, err := process.ReadProcessorMessage(r)
+	cd, res, resBlobReader, err := utils.ReadProcessorMessage(r)
 	if err != nil {
 		return fmt.Errorf("unable to read processor message: %w", err)
 	}
@@ -36,7 +37,7 @@ func (p *resourceLabeler) Process(ctx context.Context, r io.Reader, w io.Writer)
 
 	res.Labels = append(res.Labels, p.labels...)
 
-	if err := process.WriteProcessorMessage(*cd, res, resBlobReader, w); err != nil {
+	if err := utils.WriteProcessorMessage(*cd, res, resBlobReader, w); err != nil {
 		return fmt.Errorf("unable to write processor message: %w", err)
 	}
